integration/requirements.txt
requests==2.25.0
pytest==6.2.4
pyyaml==5.4
